requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==35.0.0
Deprecated==1.2.13
Django==3.2.7
django-core==1.4.1
django-rest-auth==0.9.5
djangorestframework==3.12.4
gcloud==0.18.3
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.2
jwcrypto==1.0
oauth2client==4.1.3
pkg-resources==0.0.0
protobuf==4.0.0rc2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.4
pyparsing==2.4.7
Pyrebase4==4.5.0
python-jwt==3.3.0
pytz==2021.1
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.2
urllib3==1.26.7
wrapt==1.12.1
